backend/requirements.txt
mysql-connector-python~=8.0.33
pydantic~=1.10.9
fastapi~=0.96.0
passlib~=1.7.4
opencv-python~=4.5.4.60
numpy~=1.24.3
insightface~=0.7.3
requests~=2.30.0